Improve handling of multiple messageIds in SmppClient
The code for setting messageId in the SmppClient class has been updated to accommodate multiple messageIds. Previously, the 'ReceiptedMessageId' was directly updated with the new 'messageId'. With this change, if 'ReceiptedMessageId' already contains a value, the new 'messageId' is appended to it instead of replacing it. This allows the 'ReceiptedMessageId' to hold a history of all messageIds, helpful in cases where tracking multiple message transactions is needed.