Fix MultiMessage MessaageId Retrieval Issue
This commit makes a series of changes to TextMessage and SmppClient for readability and some improved functionality.

In the TextMessage file, we've improved readability by breaking up some complex methods and reorganizing code blocks. We've factored out the SubmitSm object creation into its own method, ensuring the object is consistently created whether the text message needs splitting or not.

We've also added some whitespace for readability and tidied up some variable names for clarity.

In the SmppClient file, we've made some tweaks to improve efficiency and clarity. First, we've moved the declaration of the 'messageId' variable closer to where it's used, removing an unneeded declaration at the start. We've changed the way the response PDU is cast for simpler readability and string concatenation for the receiptedMessageId. Lastly, typo in the 'CreateDefaultResponce' method has been corrected to 'CreateDefaultResponse'.